fix wrong declaration
git-svn-id: svn://svn.dd-wrt.com/DD-WRT@47483 52c4871e-980c-0410-b1e0-e73912ce01f8

+++ src/router/shared/broadcom.h
@@ -861,7 +861,7 @@ size_t websWrite(webs_t wp, char *fmt, ...);
 void do_ej(unsigned char method, struct mime_handler *handler, char *path, webs_t stream);
 
 FILE *getWebsFile(webs_t wp, char *path);
-FILE *_getWebsFile(webs_t wp, char *path, int *len);
+FILE *_getWebsFile(webs_t wp, char *path, size_t *len);
 
 int wfputs(char *buf, webs_t fp);
